requirements.txt
Flask
Flask-Cors
langchain
langchain-community
langchain-chroma
langchain-google-genai
google-generativeai
python-dotenv
beautifulsoup4
unstructured
gunicorn
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.2.0